+++ CHANGELOG.md
@@ -5,6 +5,7 @@ GAMSPy 1.5.1
 ------------
 - General
   - Fix the bugs in dynamic set assignments.
+  - Allow loadRecordsFromGdx to domain forward.
 - Documentation
   - Update embedding Neural Network documentation.